+++ jsx-dev-runtime.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const { Fragment, createElement } = require('./src/htmx/Jsx');
+
+function jsxDEV(type, properties) {
+    return createElement(type, properties);
+}
+
+module.exports = { Fragment, jsxDEV };